initial skeleton - partially working
