+++ server/www/templates/collection-selector/collection-selector.html
@@ -1,0 +1,29 @@
+<fieldset
+    dropdown auto-close="outsideClick" 
+    class="custom-fieldset init"
+    >
+    <legend data-toggle="dropdown-menu" class="dropdown-trigger init" dropdown-toggle>
+        <span class="icon-left fa-th" translate>global_filter.collections.collections</span>
+    </legend>
+    <div class="dropdown-menu init" dropdown-menu>
+        <!-- div class="form-field search bar">
+            <label for="search-collections" class="nodisplay" translate>global_filter.collections.search_collections</label>
+            <input id="search-collections" placeholder="Search Collections" type="search">
+        </div -->
+        <div class="fieldset-results" ng-repeat="collection in editableCollections">
+            <div class="form-field checkbox">
+                <input id="collection" type="checkbox" ng-checked="postInCollection(collection)" ng-click="toggleCollection(collection)">
+                <label for="collection">{{collection.name}}</label>
+            </div>
+        </div>
+        <div class="form-field bar">
+            <a class="button button-link init icon-left fa-plus" ng-show="!showNewCollectionInput" ng-click="toggleCreateCollection()" translate>
+                set.create_collection
+            </a>
+            <div ng-show="showNewCollectionInput">
+                <input id="create-collection" ng-model="newCollection" placeholder="Create Collection" type="text">
+                <button type="button" ng-click="createNewCollection(newCollection)" translate>post.create</button>
+            </div>
+        </div>
+    </div>
+</fieldset>
